ref(browser): Remove backend
