+++ packages/browser/src/client.ts
@@ -1,10 +1,37 @@
 import { BaseClient, Scope, SDK_VERSION } from '@sentry/core';
-import { Event, EventHint } from '@sentry/types';
-import { getGlobalObject, logger } from '@sentry/utils';
+import { Event, EventHint, Options, SeverityLevel, Transport } from '@sentry/types';
+import { getGlobalObject, logger, supportsFetch } from '@sentry/utils';
 
-import { BrowserBackend, BrowserOptions } from './backend';
+import { eventFromException, eventFromMessage } from './eventbuilder';
 import { injectReportDialog, ReportDialogOptions } from './helpers';
 import { Breadcrumbs } from './integrations';
+import { FetchTransport, XHRTransport } from './transports';
+
+/**
+ * Configuration options for the Sentry Browser SDK.
+ * @see BrowserClient for more information.
+ */
+export interface BrowserOptions extends Options {
+  /**
+   * A pattern for error URLs which should exclusively be sent to Sentry.
+   * This is the opposite of {@link Options.denyUrls}.
+   * By default, all errors will be sent.
+   */
+  allowUrls?: Array<string | RegExp>;
+
+  /**
+   * A pattern for error URLs which should not be sent to Sentry.
+   * To allow certain errors instead, use {@link Options.allowUrls}.
+   * By default, all errors will be sent.
+   */
+  denyUrls?: Array<string | RegExp>;
+
+  /** @deprecated use {@link Options.allowUrls} instead. */
+  whitelistUrls?: Array<string | RegExp>;
+
+  /** @deprecated use {@link Options.denyUrls} instead. */
+  blacklistUrls?: Array<string | RegExp>;
+}
 
 /**
  * The Sentry Browser SDK Client.
@@ -12,7 +39,7 @@ import { Breadcrumbs } from './integrations';
  * @see BrowserOptions for documentation on configuration options.
  * @see SentryClient for usage documentation.
  */
-export class BrowserClient extends BaseClient<BrowserBackend, BrowserOptions> {
+export class BrowserClient extends BaseClient<BrowserOptions> {
   /**
    * Creates a new Browser SDK instance.
    *
@@ -31,7 +58,7 @@ export class BrowserClient extends BaseClient<BrowserBackend, BrowserOptions> {
       version: SDK_VERSION,
     };
 
-    super(BrowserBackend, options);
+    super(options);
   }
 
   /**
@@ -73,6 +100,47 @@ export class BrowserClient extends BaseClient<BrowserBackend, BrowserOptions> {
     if (integration) {
       integration.addSentryBreadcrumb(event);
     }
-    super._sendEvent(event);
+    super.sendEvent(event);
+  }
+
+  /**
+   * @inheritDoc
+   */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
+  protected _eventFromException(exception: any, hint?: EventHint): PromiseLike<Event> {
+    return eventFromException(this._options, exception, hint);
+  }
+
+  /**
+   * @inheritDoc
+   */
+  protected _eventFromMessage(message: string, level: SeverityLevel = 'info', hint?: EventHint): PromiseLike<Event> {
+    return eventFromMessage(this._options, message, level, hint);
+  }
+
+  /**
+   * @inheritDoc
+   */
+  protected _setupTransport(): Transport {
+    if (!this._options.dsn) {
+      // We return the noop transport here in case there is no Dsn.
+      return super._setupTransport();
+    }
+
+    const transportOptions = {
+      ...this._options.transportOptions,
+      dsn: this._options.dsn,
+      tunnel: this._options.tunnel,
+      sendClientReports: this._options.sendClientReports,
+      _metadata: this._options._metadata,
+    };
+
+    if (this._options.transport) {
+      return new this._options.transport(transportOptions);
+    }
+    if (supportsFetch()) {
+      return new FetchTransport(transportOptions);
+    }
+    return new XHRTransport(transportOptions);
   }
 }
