+++ packages/browser/src/exports.ts
@@ -38,8 +38,7 @@ export {
   withScope,
 } from '@sentry/core';
 
-export { BrowserOptions } from './backend';
-export { BrowserClient } from './client';
+export { BrowserClient, BrowserOptions } from './client';
 export { injectReportDialog, ReportDialogOptions } from './helpers';
 export { eventFromException, eventFromMessage } from './eventbuilder';
 export { defaultIntegrations, forceLoad, init, lastEventId, onLoad, showReportDialog, flush, close, wrap } from './sdk';
